add mos-unknown-none target
update target configuration

add build.sh

.gitmodules, Cargo.lock

use mos-clang as linker

bugfix

add create_mos_targets.py

set target vendor

adapt to new sdk

fix create_mos_targets script

fix build.sh

s/a800xl/atari8/g

add nes-nrom-128 target

update README.md

use cow

memcmp bugfix

use fork of cargo 0.1.75

build docker images, closes #3

trigger workflow on push to master

Provide dylib instructions for libxml2 on MacOS

add missing platforms

c_int_definition based on target_pointer_width

update .gitmodules

+++ compiler/rustc_codegen_llvm/src/debuginfo/mod.rs
@@ -179,7 +179,7 @@ impl<'ll> DebugInfoBuilderMethods for Builder<'_, 'll, '_> {
             addr_ops.push(direct_offset.bytes() as u64);
         }
         for &offset in indirect_offsets {
-            addr_ops.push(op_deref());
+            addr_ops.push(op_deref() as u64);
             if offset.bytes() > 0 {
                 addr_ops.push(op_plus_uconst());
                 addr_ops.push(offset.bytes() as u64);
